backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
mysqlclient==2.2.0
redis==5.0.1

# Async Tasks
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API Documentation
drf-spectacular==0.26.5

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# Utilities
python-decouple==3.8
Pillow==10.1.0
requests==2.31.0
python-dateutil==2.8.2

# Development
django-debug-toolbar==4.2.0
ipython==8.17.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.38.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
